+++ xray-svn-trunk/xr_3da/xrGame/Grenade.cpp
@@ -256,7 +256,7 @@ void CGrenade::UpdateCL()
 }
 
 
-bool CGrenade::Action(s32 cmd, u32 flags) 
+bool CGrenade::Action(u16 cmd, u32 flags) 
 {
 	if(inherited::Action(cmd, flags)) return true;
