+++ xray-svn-trunk/xr_3da/xrGame/WeaponKnife.h
@@ -64,7 +64,7 @@ public:
 	virtual void		FireStart						();
 
 
-	virtual bool		Action							(s32 cmd, u32 flags);
+	virtual bool		Action							(u16 cmd, u32 flags);
 	virtual void	onMovementChanged	(ACTOR_DEFS::EMoveCommand cmd);
 	virtual void		StartIdleAnim					();
 	virtual void		GetBriefInfo					(xr_string& str_name, xr_string& icon_sect_name, xr_string& str_count);
