+++ xray-svn-trunk/xr_3da/xrGame/WeaponMagazinedWGrenade.cpp
@@ -246,7 +246,7 @@ void  CWeaponMagazinedWGrenade::PerformSwitchGL()
 	}
 }
 
-bool CWeaponMagazinedWGrenade::Action(s32 cmd, u32 flags) 
+bool CWeaponMagazinedWGrenade::Action(u16 cmd, u32 flags) 
 {
 	if(inherited::Action(cmd, flags)) return true;
